+++ Assets/Script/PlayerLogic/PlayerStatManagement.cs
@@ -13,7 +13,7 @@ public class PlayerStatManagement : StatManagement
         healthBar.SetMaxHealth(maxHealth);
     }
 
-    public override void takeDamage(int damage)
+    public override void takeDamage(float damage)
     {
         base.takeDamage(damage);
         healthBar.SetHealth(currentHealth);
